Adressing warnings issue and implementing the  independent intersection test per region spheres

+++ libraries/workload/src/workload/View.h
@@ -38,7 +38,7 @@ public:
     float maxRadius{ FLT_MAX };
 
     // Fov stores the half field of view angle, and tan/cos/sin ready to go, default is fov of 90deg
-    glm::vec4 fov_halfAngle_tan_cos_sin { M_PI_4, 1.0f, M_SQRT2 * 0.5f, M_SQRT2 * 0.5f };
+    glm::vec4 fov_halfAngle_tan_cos_sin { (float) M_PI_4, 1.0f, (float) (M_SQRT2 * 0.5), (float) (M_SQRT2 * 0.5) };
 
     // Origin position 
     glm::vec3 origin{ 0.0f };
